+++ server/resources/views/documents/setting.blade.php
@@ -12,14 +12,16 @@
     <hr>
     @foreach($documents as $document)
         <div class="row">
-            <div class="col-md-8 text-center">
+            <div class="col-md-9 text-center">
                 <a href="documents/view/{{$document->id}}">
                     <strong>{{$document->document_name}}</strong>
                 </a>
             </div>
-            <div class="col-md-4 text-center">
+            <div class="col-md-3 text-center">
+                @if (Auth::user()->level >= 99)
+                    <a href="documents/delete/{{$document->id}}" class="btn btn-danger">Delete</a>
+                @endif
                 <a href="documents/edit/{{$document->id}}" class="btn btn-primary">Edit</a>
-                <a href="documents/delete/{{$document->id}}" class="btn btn-danger">Delete</a>
             </div>
         </div>
         <hr>
